migrate utils and move trackDappVisitedEvent to utils folder
