+++ app/components/Views/AccountConnect/AccountConnect.tsx
@@ -54,7 +54,7 @@ import AccountConnectSingleSelector from './AccountConnectSingleSelector';
 import AccountConnectMultiSelector from './AccountConnectMultiSelector';
 import useFavicon from '../../hooks/useFavicon/useFavicon';
 import URLParse from 'url-parse';
-import { trackDappVisitedEvent } from '../../../analytics';
+import { trackDappVisitedEvent } from '../../../util/metrics';
 
 const AccountConnect = (props: AccountConnectProps) => {
   const Engine = UntypedEngine as any;
