+++ app/util/metrics/index.ts
@@ -1,7 +1,9 @@
 import DeviceAnalyticsMetaData from './DeviceAnalyticsMetaData/generateDeviceAnalyticsMetaData';
 import UserSettingsAnalyticsMetaData from './UserSettingsAnalyticsMetaData/generateUserProfileAnalyticsMetaData';
 import TrackAfterInteractions from './TrackAfterInteraction/trackAfterInteractions';
+import trackDappVisitedEvent from './trackDappVisitedEvent';
 
 export default DeviceAnalyticsMetaData;
 export { UserSettingsAnalyticsMetaData };
 export { TrackAfterInteractions };
+export { trackDappVisitedEvent };
